+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 > :warning: represents a breaking change
 
+## [1.4.3] - 2021-03-31
+### Added
+- More unit tests
+
+### Changed
+- Enforce widget types (`SimpleStack` is now a `Stack`; `SafeScaffold` is now a `SafeArea`; 
+  `OnDemandListView` is now a `ListView`; `SimpleVisibility` is now a `Visibility`; `SplashScreen`
+  is now a `FutureBuilder`)
+
+### Fixed
+- :warning: `floatingActionBar` parameter of `SafeScaffold` is renamed to `floatingActionButton`
+
+
 ## [1.4.2] - 2021-03-31
 ### Added
 - `FontStyle` attribute to `StyledText`